+++ mobile/android/android-components/components/support/ktx/src/test/java/mozilla/components/support/ktx/android/view/ActivityTest.kt
@@ -12,16 +12,15 @@ import android.view.Window
 import android.view.WindowInsets
 import android.view.WindowManager
 import androidx.core.view.WindowInsetsCompat
+import androidx.core.view.WindowInsetsControllerCompat
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import mozilla.components.support.test.any
 import mozilla.components.support.test.argumentCaptor
 import mozilla.components.support.test.mock
-import org.junit.Assert
 import org.junit.Assert.assertEquals
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
-import org.mockito.Mockito.anyInt
 import org.mockito.Mockito.doReturn
 import org.mockito.Mockito.never
 import org.mockito.Mockito.times
@@ -35,45 +34,37 @@ class ActivityTest {
 
     private lateinit var activity: Activity
     private lateinit var window: Window
-    private lateinit var layoutParams: WindowManager.LayoutParams
     private lateinit var decorView: View
     private lateinit var viewTreeObserver: ViewTreeObserver
-    private lateinit var windowInsetsCompat: WindowInsetsCompat
     private lateinit var windowInsets: WindowInsets
+    private lateinit var insetsController: WindowInsetsControllerCompat
+    private lateinit var layoutParams: WindowManager.LayoutParams
 
     @Before
     fun setup() {
         activity = mock()
         window = mock()
-        layoutParams = WindowManager.LayoutParams()
         decorView = mock()
         viewTreeObserver = mock()
-        windowInsetsCompat = mock()
         windowInsets = mock()
+        insetsController = mock()
+        layoutParams = WindowManager.LayoutParams()
 
         `when`(activity.window).thenReturn(window)
         `when`(window.decorView).thenReturn(decorView)
         `when`(window.decorView.viewTreeObserver).thenReturn(viewTreeObserver)
-        `when`(window.attributes).thenReturn(layoutParams)
-        `when`(windowInsetsCompat.toWindowInsets()).thenReturn(windowInsets)
         `when`(window.decorView.onApplyWindowInsets(windowInsets)).thenReturn(windowInsets)
-    }
-
-    @Test
-    fun `check enterToImmersiveMode sets the correct flags`() {
-        activity.enterToImmersiveMode()
-
-        // verify entering immersive mode
-        verify(decorView).systemUiVisibility = View.SYSTEM_UI_FLAG_FULLSCREEN
-        verify(decorView).systemUiVisibility = View.SYSTEM_UI_FLAG_HIDE_NAVIGATION
-        // verify that the immersive mode restoration is set as expected
-        verify(window.decorView).setOnApplyWindowInsetsListener(any())
+        `when`(window.attributes).thenReturn(layoutParams)
     }
 
     @Test
     @Config(sdk = [Build.VERSION_CODES.P])
-    fun `GIVEN Android version P WHEN setAsImmersive is called THEN notch flags are set`() {
-        activity.setAsImmersive()
+    fun `GIVEN Android version P WHEN enterImmersiveMode is called THEN systems bars are hidden, inset listener is set and notch flags are set to extend view into notch area`() {
+        activity.enterImmersiveMode(insetsController)
+
+        verify(insetsController).hide(WindowInsetsCompat.Type.systemBars())
+        verify(insetsController).systemBarsBehavior = WindowInsetsControllerCompat.BEHAVIOR_SHOW_TRANSIENT_BARS_BY_SWIPE
+        verify(window.decorView).setOnApplyWindowInsetsListener(any())
 
         verify(window).setFlags(WindowManager.LayoutParams.FLAG_LAYOUT_NO_LIMITS, WindowManager.LayoutParams.FLAG_LAYOUT_NO_LIMITS)
         assertEquals(WindowManager.LayoutParams.LAYOUT_IN_DISPLAY_CUTOUT_MODE_SHORT_EDGES, layoutParams.layoutInDisplayCutoutMode)
@@ -81,77 +72,72 @@ class ActivityTest {
 
     @Test
     @Config(sdk = [Build.VERSION_CODES.O_MR1])
-    fun `GIVEN Android version O_MR1 WHEN setAsImmersive is called THEN notch flags are not being set`() {
-        activity.setAsImmersive()
+    fun `GIVEN Android version O_MR1 WHEN enterImmersiveMode is called THEN systems bars are hidden, inset listener is set and notch flags are not being set`() {
+        activity.enterImmersiveMode(insetsController)
+
+        verify(insetsController).hide(WindowInsetsCompat.Type.systemBars())
+        verify(insetsController).systemBarsBehavior = WindowInsetsControllerCompat.BEHAVIOR_SHOW_TRANSIENT_BARS_BY_SWIPE
+        verify(window.decorView).setOnApplyWindowInsetsListener(any())
 
         verify(window, never()).setFlags(WindowManager.LayoutParams.FLAG_LAYOUT_NO_LIMITS, WindowManager.LayoutParams.FLAG_LAYOUT_NO_LIMITS)
     }
 
     @Test
-    fun `check enableImmersiveModeRestore sets insets listeners`() {
-        activity.enableImmersiveModeRestore()
-
-        verify(window.decorView).setOnApplyWindowInsetsListener(any())
-    }
-
-    @Test
-    fun `check enableImmersiveModeRestore set insets listeners have the correct behavior when status bar is visible`() {
+    fun `GIVEN enterImmersiveMode was called WHEN window insets are changed THEN insetsController hides system bars and sets bars behaviour again`() {
         val insetListenerCaptor = argumentCaptor<View.OnApplyWindowInsetsListener>()
-
         doReturn(30).`when`(windowInsets).systemWindowInsetTop
 
-        activity.enableImmersiveModeRestore()
-        verify(window.decorView).setOnApplyWindowInsetsListener(insetListenerCaptor.capture())
+        activity.enterImmersiveMode(insetsController)
 
+        verify(insetsController, times(1)).hide(WindowInsetsCompat.Type.systemBars())
+        verify(insetsController, times(1)).systemBarsBehavior = WindowInsetsControllerCompat.BEHAVIOR_SHOW_TRANSIENT_BARS_BY_SWIPE
+
+        verify(window.decorView).setOnApplyWindowInsetsListener(insetListenerCaptor.capture())
         insetListenerCaptor.value.onApplyWindowInsets(window.decorView, windowInsets)
 
-        // Cannot test if "setAsImmersive()" was called it being an extension function but we can check the effect of that call.
-        verify(decorView).systemUiVisibility = View.SYSTEM_UI_FLAG_FULLSCREEN
-        verify(decorView).systemUiVisibility = View.SYSTEM_UI_FLAG_HIDE_NAVIGATION
+        verify(insetsController, times(2)).hide(WindowInsetsCompat.Type.systemBars())
+        verify(insetsController, times(2)).systemBarsBehavior = WindowInsetsControllerCompat.BEHAVIOR_SHOW_TRANSIENT_BARS_BY_SWIPE
     }
 
     @Test
-    fun `check enableImmersiveModeRestore set insets listeners have the correct behavior when status bar is NOT visible`() {
+    fun `GIVEN enterImmersiveMode was called WHEN window insets are not changed THEN insetsController does nothing`() {
         val insetListenerCaptor = argumentCaptor<View.OnApplyWindowInsetsListener>()
-
         doReturn(0).`when`(windowInsets).systemWindowInsetTop
 
-        activity.enableImmersiveModeRestore()
-        verify(window.decorView).setOnApplyWindowInsetsListener(insetListenerCaptor.capture())
+        activity.enterImmersiveMode(insetsController)
 
+        verify(insetsController, times(1)).hide(WindowInsetsCompat.Type.systemBars())
+        verify(insetsController, times(1)).systemBarsBehavior = WindowInsetsControllerCompat.BEHAVIOR_SHOW_TRANSIENT_BARS_BY_SWIPE
+
+        verify(window.decorView).setOnApplyWindowInsetsListener(insetListenerCaptor.capture())
         insetListenerCaptor.value.onApplyWindowInsets(window.decorView, windowInsets)
 
-        // Cannot test if "setAsImmersive()" was called it being an extension function but we can check the effect of that call.
-        verify(window, never()).addFlags(anyInt())
-        verify(decorView, never()).systemUiVisibility = anyInt()
-        verify(decorView, never()).systemUiVisibility = anyInt()
+        verify(insetsController, times(1)).hide(WindowInsetsCompat.Type.systemBars())
+        verify(insetsController, times(1)).systemBarsBehavior = WindowInsetsControllerCompat.BEHAVIOR_SHOW_TRANSIENT_BARS_BY_SWIPE
     }
 
     @Test
-    @Config(sdk = [Build.VERSION_CODES.O_MR1])
-    fun `check exitImmersiveMode sets the correct flags devices with an Android version prior to P`() {
-        activity.exitImmersiveMode()
+    fun `WHEN exitImmersiveMode is called THEN insetsController shows system bars and OnApplyWindowInsetsListener is cleared`() {
+        activity.exitImmersiveMode(insetsController)
 
-        verify(decorView, times(2)).systemUiVisibility = View.SYSTEM_UI_FLAG_VISIBLE
-        verify(decorView).setOnApplyWindowInsetsListener(null)
+        verify(insetsController).show(WindowInsetsCompat.Type.systemBars())
+        verify(window.decorView).setOnApplyWindowInsetsListener(null)
     }
 
     @Test
     @Config(sdk = [Build.VERSION_CODES.P])
-    fun `check exitImmersiveMode sets the correct flags on devices with Android P`() {
+    fun `GIVEN Android version P WHEN exitImmersiveMode is called THEN notch flags are reset to defaults`() {
         activity.exitImmersiveMode()
 
-        verify(decorView, times(2)).systemUiVisibility = View.SYSTEM_UI_FLAG_VISIBLE
-        verify(decorView).setOnApplyWindowInsetsListener(null)
-
         verify(window).clearFlags(WindowManager.LayoutParams.FLAG_LAYOUT_NO_LIMITS)
-        Assert.assertEquals(WindowManager.LayoutParams.LAYOUT_IN_DISPLAY_CUTOUT_MODE_DEFAULT, layoutParams.layoutInDisplayCutoutMode)
+        assertEquals(WindowManager.LayoutParams.LAYOUT_IN_DISPLAY_CUTOUT_MODE_DEFAULT, layoutParams.layoutInDisplayCutoutMode)
     }
 
     @Test
-    fun `check exitImmersiveMode correctly cleanups the insets listeners`() {
+    @Config(sdk = [Build.VERSION_CODES.O_MR1])
+    fun `GIVEN Android version O_MR1 WHEN exitImmersiveMode is called THEN notch flags were not being set`() {
         activity.exitImmersiveMode()
 
-        verify(decorView).setOnApplyWindowInsetsListener(null)
+        verify(window, never()).setFlags(WindowManager.LayoutParams.FLAG_LAYOUT_NO_LIMITS, WindowManager.LayoutParams.FLAG_LAYOUT_NO_LIMITS)
     }
 }